refactored Test structure
